+++ View/Controls/ArtistGridCardControl.xaml.cs
@@ -1,0 +1,19 @@
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using WinUIMusicPlayer.Model;
+
+namespace WinUIMusicPlayer.View.Controls;
+
+public sealed partial class ArtistGridCardControl : UserControl
+{
+    public static readonly DependencyProperty MusicProperty = DependencyProperty.Register(
+        nameof(Music), typeof(Music), typeof(ArtistGridCardControl), new PropertyMetadata(null));
+
+    public Music? Music
+    {
+        get => (Music?)GetValue(MusicProperty);
+        set => SetValue(MusicProperty, value);
+    }
+
+    public ArtistGridCardControl() => InitializeComponent();
+}